# docs/swagger.yaml
basePath: /api
definitions:
  Attachment:
    properties:
      filename:
        type: string
      filetype:
        type: string
      url:
        type: string
    type: object
  BanResponse:
    properties:
      id:
        type: string
      image:
        type: string
      username:
        type: string
    type: object
...
  Member:
    properties:
      color:
        type: string
      createdAt:
        type: string
      id:
        type: string
      image:
        type: string
      isFriend:
        type: boolean
      isOnline:
        type: boolean
      nickname:
        type: string
      updatedAt:
        type: string
      username:
        type: string
    type: object
  MemberRequest:
    properties:
      memberId:
        type: string
    type: object
  MemberSettings:
    properties:
      color:
        type: string
      nickname:
        type: string
    type: object
  Message:
    properties:
      attachment:
        $ref: '#/definitions/Attachment'
      createdAt:
        type: string
      id:
        type: string
      text:
        type: string
      updatedAt:
        type: string
      user:
        $ref: '#/definitions/Member'
    type: object
  MessageRequest:
    properties:
      file:
        description: image/* or audio/*
        format: binary
        type: string
      text:
        description: Maximum 2000 characters
        type: string
    type: object
...
  SuccessResponse:
    properties:
      success:
        description: Only returns true, not an json object
        type: boolean
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Valkyrie REST API Specs. This service uses sessions for authentication
  license:
    name: Apache 2.0
  title: Valkyrie API
  version: "1.0"
paths:
...
  /guilds/{guildId}/bans:
    delete:
      parameters:
      - description: Guild ID
        in: path
        name: guildId
        required: true
        type: string
      - description: Member ID
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/MemberRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/SuccessResponse'
            type: array
      summary: Unban Member
      tags:
      - Members
    get:
      parameters:
      - description: Guild ID
        in: path
        name: guildId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/BanResponse'
            type: array
      summary: Get Guild Ban list
      tags:
      - Members
    post:
      parameters:
      - description: Guild ID
        in: path
        name: guildId
        required: true
        type: string
      - description: Member ID
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/MemberRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/SuccessResponse'
            type: array
      summary: Ban Member
      tags:
      - Members
...
  /guilds/{guildId}/kick:
    post:
      parameters:
      - description: Guild ID
        in: path
        name: guildId
        required: true
        type: string
      - description: Member ID
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/MemberRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/SuccessResponse'
            type: array
      summary: Kick Member
      tags:
      - Members
  /guilds/{guildId}/member:
    get:
      parameters:
      - description: Guild ID
        in: path
        name: guildId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/MemberSettings'
      summary: Get Member Settings
      tags:
      - Members
    put:
      parameters:
      - description: Edit Member
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/MemberSettings'
      - description: Guild ID
        in: path
        name: guildId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/SuccessResponse'
      summary: Edit Member Settings
      tags:
      - Members
...
  /messages/{channelId}:
    get:
      parameters:
      - description: Channel ID
        in: path
        name: channelId
        required: true
        type: string
      - description: Cursor Pagination using the createdAt field
        in: query
        name: cursor
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/Message'
            type: array
      summary: Get Channel Messages
      tags:
      - Messages
    post:
      parameters:
      - description: Channel ID
        in: path
        name: channelId
        required: true
        type: string
      - description: Create Message
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/MessageRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/SuccessResponse'
      summary: Create Messages
      tags:
      - Messages
  /messages/{messageId}:
    delete:
      parameters:
      - description: Message ID
        in: path
        name: messageId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/SuccessResponse'
      summary: Delete Messages
      tags:
      - Messages
    put:
      parameters:
      - description: Message ID
        in: path
        name: messageId
        required: true
        type: string
      - description: Edit Message
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/MessageRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/SuccessResponse'
      summary: Edit Messages
      tags:
      - Messages
swagger: "2.0"